Moved search bar to a component and changed the conditions
